Move common file operations into file_utils.py
But don't break API.
