+++ snapcraft/__init__.py
@@ -237,4 +237,5 @@ from snapcraft._store import (                      # noqa
 from snapcraft import common                        # noqa
 from snapcraft import plugins                       # noqa
 from snapcraft import sources                       # noqa
+from snapcraft import file_utils                    # noqa
 from snapcraft.internal import repo                 # noqa
