+++ snapcraft/internal/common.py
@@ -163,75 +163,6 @@ def reset_env():
     env = []
 
 
-def link_or_copy(source, destination, follow_symlinks=False):
-    """Hard-link source and destination files. Copy if it fails to link.
-
-    Hard-linking may fail (e.g. a cross-device link, or permission denied), so
-    as a backup plan we just copy it.
-
-    :param str source: The source to which destination will be linked.
-    :param str destination: The destination to be linked to source.
-    :param bool follow_symlinks: Whether or not symlinks should be followed.
-    """
-
-    try:
-        # Note that follow_symlinks doesn't seem to work for os.link, so we'll
-        # implement this logic ourselves using realpath.
-        source_path = source
-        if follow_symlinks:
-            source_path = os.path.realpath(source)
-
-        # Setting follow_symlinks=False in case this bug is ever fixed
-        # upstream-- we want this function to continue supporting NOT following
-        # symlinks.
-        os.link(source_path, destination, follow_symlinks=False)
-    except OSError:
-        shutil.copy2(source, destination, follow_symlinks=follow_symlinks)
-        uid = os.stat(source, follow_symlinks=follow_symlinks).st_uid
-        gid = os.stat(source, follow_symlinks=follow_symlinks).st_gid
-        try:
-            os.chown(destination, uid, gid, follow_symlinks=follow_symlinks)
-        except PermissionError as e:
-            logger.warning('unable to chown {destination}: {error}'.format(
-                destination=destination, error=e))
-
-
-def replace_in_file(directory, file_pattern, search_pattern, replacement):
-    """Searches and replaces patterns that match a file pattern.
-    :param str directory: The directory to look for files.
-    :param str file_pattern: The file pattern to match inside directory.
-    :param search_pattern: A re.compile'd pattern to search for within
-                           matching files.
-    :param str replacement: The string to replace the matching search_pattern
-                            with.
-    """
-    for root, directories, files in os.walk(directory):
-        for file_name in files:
-            if file_pattern.match(file_name):
-                _search_and_replace_contents(os.path.join(root, file_name),
-                                             search_pattern, replacement)
-
-
-def _search_and_replace_contents(file_path, search_pattern, replacement):
-    # Don't bother trying to rewrite a symlink. It's either invalid or the
-    # linked file will be rewritten on its own.
-    if os.path.islink(file_path):
-        return
-
-    with open(file_path, 'r+') as f:
-        try:
-            original = f.read()
-        except UnicodeDecodeError:
-            # This was probably a binary file. Skip it.
-            return
-
-        replaced = search_pattern.sub(replacement, original)
-        if replaced != original:
-            f.seek(0)
-            f.truncate()
-            f.write(replaced)
-
-
 def get_terminal_width(max_width=MAX_CHARACTERS_WRAP):
     if os.isatty(sys.stdout.fileno()):
         width = shutil.get_terminal_size().columns
