+++ snapcraft/plugins/copy.py
@@ -37,7 +37,6 @@ Additionally, this plugin uses the following plugin-specific keywords:
 import logging
 import os
 import glob
-import shutil
 
 import snapcraft
 
@@ -123,31 +122,8 @@ def _recursively_link(source, destination, boundary):
             raise NotADirectoryError(
                 'Cannot overwrite non-directory {!r} with directory '
                 '{!r}'.format(destination, source))
-        _linktree(source, destination, boundary)
+        snapcraft.file_utils.link_or_copy_tree(
+            source, destination,
+            copy_function=lambda src, dst: _link_or_copy(src, dst, boundary))
     else:
         _link_or_copy(source, destination, boundary)
-
-
-def _create_similar_directory(source, destination):
-    os.makedirs(destination, exist_ok=True)
-    shutil.copystat(source, destination, follow_symlinks=False)
-
-
-def _linktree(source_tree, destination_tree, boundary):
-    if not os.path.isdir(destination_tree):
-        _create_similar_directory(source_tree, destination_tree)
-
-    for root, directories, files in os.walk(source_tree):
-        for directory in directories:
-            source = os.path.join(root, directory)
-            destination = os.path.join(
-                destination_tree, os.path.relpath(source, source_tree))
-
-            _create_similar_directory(source, destination)
-
-        for file_name in files:
-            source = os.path.join(root, file_name)
-            destination = os.path.join(
-                destination_tree, os.path.relpath(source, source_tree))
-
-            _link_or_copy(source, destination, boundary)
